moved orphaned files config into DownloadCleaner namespace
