+++ code/backend/Cleanuparr.Persistence/DataContext.cs
@@ -9,7 +9,6 @@ using Cleanuparr.Persistence.Models.Configuration.MalwareBlocker;
 using Cleanuparr.Persistence.Models.Configuration.Notification;
 using Cleanuparr.Persistence.Models.Configuration.QueueCleaner;
 using Cleanuparr.Persistence.Models.Configuration.BlacklistSync;
-using Cleanuparr.Persistence.Models.Configuration.OrphanedFilesCleanup;
 using Cleanuparr.Persistence.Models.Configuration.Seeker;
 
 using Cleanuparr.Persistence.Models.State;
